+++ features/step_definitions/general_steps/general_steps.rb
@@ -55,6 +55,7 @@ Given("I confirm my email address") do
 end
 
 Given("I authorize the attribute release") do
+  page.has_button? "Allow" #waits for allow button to show
   Consent.new.allow
 end